Add a PoC Excel exporter to the lib as well
After struggeling too much with paths and imports I gave up and added a
copy of the Excel creator in the lib folder. This is not nice but it is
good enough for a proof of concept. Template excel has also been put in
the same folder and contains a full set of races ranging from 25m to
1500m.