Improve handling of personal bests
Several improvements in this patch:
- Add safety when trying to get non-existing personal bests
- Handle relays so that the races are no longer ignored
- Make it clear when a swimmer is ignored

Instead of just trying to fetch a non-existing key from a dict, verify
its existence first. If it is not present we handle it gracefully
instead allowing for the second one to be fetched. It can happen that a
swimmer only has done a race in long or short course. When so fetch the
existing one.